+++ src/main/java/com/swproj/SWProject/roommangement/service/impl/FloorServiceImpl.java
@@ -6,11 +6,14 @@ import com.swproj.SWProject.roommangement.dto.res.floor.GetFloorByIdResDTO;
 import com.swproj.SWProject.roommangement.dto.res.floor.GetFloorResDTO;
 import com.swproj.SWProject.roommangement.entity.CollegeEntity;
 import com.swproj.SWProject.roommangement.entity.FloorEntity;
+import com.swproj.SWProject.roommangement.entity.RoomEntity;
 import com.swproj.SWProject.roommangement.repo.CollegeRepo;
 import com.swproj.SWProject.roommangement.repo.FloorRepo;
+import com.swproj.SWProject.roommangement.repo.RoomRepo;
 import com.swproj.SWProject.roommangement.service.FloorService;
 import com.swproj.SWProject.roommangement.service.impl.mapper.floor.FloorEntityToGetAllFloorMapper;
 import com.swproj.SWProject.roommangement.service.impl.mapper.floor.FloorEntityToGetFloorMapper;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -23,6 +26,7 @@ public class FloorServiceImpl implements FloorService {
     private final FloorEntityToGetFloorMapper floorEntityToGetFloorMapper;
     private final FloorEntityToGetAllFloorMapper floorEntityToGetAllFloorMapper;
     private final CollegeRepo collegeRepo;
+    private final RoomRepo roomRepo;
     @Override
     public void createFloor(CreateFloorReqDTO createFloorReqDTO) {
         FloorEntity floorEntity = new FloorEntity();
@@ -40,7 +44,18 @@ public class FloorServiceImpl implements FloorService {
 
     @Override
     public void removeFloor(long id) {
-        floorRepo.deleteById(id);
+        var floorOpt = floorRepo.findById(id);
+        if (floorOpt.isPresent()) {
+            List<Long> roomsId = floorOpt.get().getRooms().stream()
+                    .map(RoomEntity::getId)
+                    .toList();
+            for (Long roomId : roomsId) {
+                roomRepo.deleteById(roomId);
+            }
+            floorRepo.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Floor with ID " + id + " not found");
+        }
     }
 
     @Override
